+++ src/app/admin/package/packages/packages.component.ts
@@ -1,48 +1,76 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Subject } from 'rxjs';
 import { AdminService } from '../../admin.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-packages',
   templateUrl: './packages.component.html',
   styleUrls: ['./packages.component.scss']
 })
-export class PackagesComponent implements OnInit,OnDestroy {
-  dtOptions : DataTables.Settings = {};
-  dtTrigger : Subject<any>        = new Subject<any>();
-  packages  : any                 = [];
-  dtInitial : boolean             = false
+
+export class PackagesComponent implements OnInit, OnDestroy {
+
+  dtOptions: DataTables.Settings = {};
+  dtTrigger: Subject<any> = new Subject<any>();
+  packages: any = [];
+  dtInitial: boolean = false
+  showPackageList=false;
+
   constructor(
-    private as:AdminService
+    private as: AdminService,
+    private router: Router
   ) { }
 
   ngOnInit(): void {
     this.getPackages();
   }
 
-  getPackages(){
-    this.as.getAllPackages().subscribe((res)=>{
-      this.packages = res.packages;
-      this.dtTrigger.next();
-    },
-    (error)=>{
-      console.log(error);
-    })
+  getPackages() {
+    /* we cannot initialize admintoken globally becuase in case we change user 
+    token from browser(Apllication) then our method will not get new token and 
+    it always get global token,, */ 
+    let admintoken = localStorage.getItem("adminToken");
+    //in case we delete token from browser(Application)
+    if (admintoken != null) {
+      this.as.getAllPackages().subscribe((res) => {
+        if(res)
+        {
+        this.packages = res.packages;
+        this.dtTrigger.next();
+        this.showPackageList=true;
+        }
+      },
+        (error) => {
+          console.log(error);
+        })
+    }
+    else {
+      this.router.navigateByUrl('admin/login');
+    }
   }
 
-  deletePackage(id){
-    this.as.deletePackageById(id).subscribe((res)=>{
-      console.log(res);
-      this.getPackages();
-    },
-    (error)=>{
-      console.log(error);
-    })
+  deletePackage(id) {
+    /* we cannot initialize admintoken globally becuase in case we change user 
+    token from browser(Apllication) then our method will not get new token and 
+    it always get global token,, */ 
+    let admintoken = localStorage.getItem("adminToken");
+    //in case we delete token from browser(Application)
+    if (admintoken != null) {
+      this.as.deletePackageById(id).subscribe((res) => {
+        this.getPackages();
+      },
+        (error) => {
+          console.log(error);
+        })
+    }
+    else {
+      this.router.navigateByUrl('admin/login');
+    }
   }
 
   ngOnDestroy(): void {
     // Do not forget to unsubscribe the event
     this.dtTrigger.unsubscribe();
   }
-
 }
